+++ src/Data/Sequence/Lens.hs
@@ -12,8 +12,7 @@
 --
 ----------------------------------------------------------------------------
 module Data.Sequence.Lens
-  ( ordinal
-  , viewL, viewR
+  ( viewL, viewR
   , _head, _tail
   , _last, _init
   , sliced, slicedTo, slicedFrom
@@ -29,21 +28,6 @@ import Data.Sequence as Seq
 -- >>> import Debug.SimpleReflect.Vars as Vars hiding (f,g)
 -- >>> let f :: Expr -> Expr; f = Debug.SimpleReflect.Vars.f
 -- >>> let g :: Expr -> Expr; g = Debug.SimpleReflect.Vars.g
-
--- | A 'Lens' that can access the @n@th element of a 'Seq'.
---
--- >>> Seq.fromList [a,b,c,d] & ordinal 2 %~ f
--- fromList [a,b,f c,d]
---
--- >>> Seq.fromList [a,b,c,d] & ordinal 2 .~ e
--- fromList [a,b,e,d]
---
--- >>> Seq.fromList [a,b,c,d] ^. ordinal 2
--- c
---
--- *NB:* This is only a legal lens if there is already such an element!
-ordinal :: Int -> SimpleIndexedLens Int (Seq a) a
-ordinal i = indexed $ \ f m -> f i (index m i) <&> \a -> update i a m
 
 -- * Sequence isomorphisms