+++ src/Data/Vector/Lens.hs
@@ -41,7 +41,7 @@ import Prelude hiding ((++), length, null, head, tail, init, last, map, reverse)
 import Data.List (nub)
 import Data.Monoid
 
--- | A 'Traversal' reading and writing to the 'head' of a /non-empty/ 'Vector'
+-- | A 'Traversal' reading and writing to the 'head' of a 'Vector'
 --
 -- >>> Vector.fromList [1,2,3] ^? _head
 -- Just 1
@@ -57,7 +57,7 @@ _head f v
   | otherwise = f (head v) <&> \a -> v // [(0,a)]
 {-# INLINE _head #-}
 
--- | A 'Traversal' reading and writing to the 'last' element of a /non-empty/ 'Vector'
+-- | A 'Traversal' reading and writing to the 'last' element of a 'Vector'
 --
 -- >>> Vector.fromList "abcde" ^? _last
 -- Just 'e'
@@ -73,7 +73,7 @@ _last f v
   | otherwise = f (last v) <&> \a -> v // [(length v - 1, a)]
 {-# INLINE _last #-}
 
--- | A 'Traversal' reading and writing to the 'tail' of a /non-empty/ 'Vector'
+-- | A 'Traversal' reading and writing to the 'tail' of a 'Vector'
 --
 -- >>> _tail .~ Vector.fromList [3,4,5] $ Vector.fromList [1,2]
 -- fromList [1,3,4,5]
@@ -83,7 +83,7 @@ _tail f v
   | otherwise = f (tail v) <&> cons (head v)
 {-# INLINE _tail #-}
 
--- | A 'Traversal' reading and replacing all but the a 'last' element of a /non-empty/ 'Vector'
+-- | A 'Traversal' reading and replacing all but the a 'last' element of a 'Vector'
 --
 -- >>> Vector.fromList [1,2,3,4] ^? _init
 -- Just (fromList [1,2,3])