+++ CHANGELOG.markdown
@@ -1,6 +1,7 @@
 3.8
 ---
-* Moved `At` to `Control.Lens.At` and factored out `Ixed(ix)`. Deprecated `_at`.
+* Moved `At(at)` and `Contains(contains)` to `Control.Lens.At` and factored out `Ixed(ix)`.
+* Deprecated `_at` and `resultAt`.
 * Removed various `ordinal` and `ix` combinators, which are subsumed by `Ixed(ix)`.
 * Changed `_head`, `_tail`, `_init` and `_last` to traversals in `Data.Vector.Lens` and `Data.Vector.Generic.Lens`.